+++ Tests/IntegrationTestProject/ProjectUnderTest/TestQualityProfile.xml
@@ -1,0 +1,31 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<profile>
+  <name>ProfileForTest</name>
+  <language>cs</language>
+  <rules>
+    <rule>
+      <repositoryKey>fxcop</repositoryKey>
+      <key>DoNotRaiseReservedExceptionTypes</key>
+      <priority>MAJOR</priority>
+      <parameters/>
+    </rule>
+    <rule>
+      <repositoryKey>fxcop</repositoryKey>
+      <key>DoNotPassLiteralsAsLocalizedParameters</key>
+      <priority>MAJOR</priority>
+      <parameters/>
+    </rule>
+    <rule>
+      <repositoryKey>csharpsquid</repositoryKey>
+      <key>S2228</key>
+      <priority>CRITICAL</priority>
+      <parameters/>
+    </rule>
+    <rule>
+      <repositoryKey>csharpsquid</repositoryKey>
+      <key>S1134</key>
+      <priority>MAJOR</priority>
+      <parameters/>
+    </rule>
+  </rules>
+</profile>
